refactor(analysis): テストファイル整理とネスティング設定 (#252)
- インラインテストを外部ファイル化（src/ フラット配置）
  - consts.rs → consts_tests.rs
  - units.rs → units_tests.rs
  - geometry/plane3_tests.rs (tests/ から移動)
  - numerics/integration_tests.rs, vector_distance_tests.rs

- テストファイルの統合・整理
  - linalg/scalar_tests.rs → abstract_types/scalar_tests.rs に統合
  - linalg/matrix/matrix3_affine_tests.rs → matrix3_tests.rs に統合
  - linalg/solver/solver_tests.rs → mod_tests.rs にリネーム

- VS Code ネスティング設定の追加・修正
  - analysis 専用パターンを追加
  - 削除したファイルのパターンを除外

- 不要ファイルの削除
  - foundation/analysis/src/mod.rs (deprecated)

テスト結果: 137件のユニットテスト + 22件のドキュメントテスト 全て合格
