+++ .vscode/settings.json
@@ -173,8 +173,30 @@
     "curve_2d.rs": "curve_2d_foundation.rs,curve_2d_transform.rs",
     "curve_3d.rs": "curve_3d_extensions.rs,curve_3d_foundation.rs,curve_3d_transform.rs",
     "surface_3d.rs": "surface_3d_foundation.rs,surface_3d_transform.rs",
+    // analysis クレート（このブランチ作業対象）
+    "consts.rs": "consts_tests.rs",
+    "units.rs": "units_tests.rs",
+    "plane3.rs": "plane3_tests.rs",
+    "integration.rs": "integration_tests.rs",
+    "vector_distance.rs": "vector_distance_tests.rs",
+    "point2.rs": "point2_tests.rs",
+    "point3.rs": "point3_tests.rs",
+    "quaternion.rs": "quaternion_tests.rs",
+    "matrix2.rs": "matrix2_tests.rs",
+    "matrix3.rs": "matrix3_tests.rs",
+    "matrix4.rs": "matrix4_tests.rs",
+    "vector.rs": "vector_tests.rs",
+    "vector2.rs": "vector2_tests.rs",
+    "vector3.rs": "vector3_tests.rs",
+    "vector4.rs": "vector4_tests.rs",
+    "cramer.rs": "cramer_tests.rs",
+    "gaussian.rs": "gaussian_tests.rs",
+    "lu.rs": "lu_tests.rs",
+    "newton.rs": "newton_tests.rs",
+    "mod.rs": "mod_tests.rs",
+    "angle.rs": "angle_tests.rs",
     // その他
-    "lib.rs": "foundation_tests.rs,*_tests.rs",
+    "lib.rs": "foundation_tests.rs",
     "transform.rs": "transform_error.rs",
     "collision.rs": "collision_error.rs",
     "intersection.rs": "intersection_error.rs",
