+++ foundation/analysis/src/linalg/solver/mod_tests.rs
