--- foundation/analysis/src/mod.rs
@@ -1,2 +1,0 @@
-﻿// This file is deprecated - use lib.rs for module organization
-// All modules are now organized in lib.rs with proper structure
